requirements.txt
Django>=4.0.0
djangorestframework>=3.12.0
djangorestframework-simplejwt>=5.0.0
django-cors-headers>=3.10.0
python-decouple>=3.5
Pillow>=8.3.0
psycopg2-binary>=2.9.0
djangorestframework-simplejwt
